Extract DbPagerTest anonymous fakes into tests/Fake/
Move the inline anonymous test doubles used by DbPagerTest into named
fake classes under tests/Fake/, matching the project convention:

- FakeSqlQuery implements SqlQueryInterface and records perPageHistory
- FakeEmptyPages implements PagesInterface wrapping an empty ArrayIterator
- FakeMediaQueryLogger implements MediaQueryLoggerInterface as a no-op

DbPagerTest now constructs DbPager with these named fakes, keeping the
test body focused on the mutation assertion.
